requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
slack_bolt==1.23.0
slack_sdk==3.36.0
Werkzeug==3.1.3
